+++ src/Adonan/BahanRotiPizzaPerPcs.java
@@ -55,7 +55,7 @@ public class BahanRotiPizzaPerPcs extends Pizza implements TotalBeratBahan {
     }
 
     @Override
-    public int getTotalBerat() {
+    public double getTotalBerat() {
         int TotalBeratPizza = (int)(tepungTerigu()+gulaPasir()+butter()+ragi()+susuBubuk()+susuCair() + esBatu() + telur());
         return beratTotalAdonanPizza;
     }
