+++ src/Adonan/BahanRotiTawarPerPcs.java
@@ -55,7 +55,7 @@ public class BahanRotiTawarPerPcs extends RotiTawar implements TotalBeratBahan{
     }
 
     @Override
-    public int getTotalBerat() {
+    public double getTotalBerat() {
         int totalBerat = (int) (tepungTerigu()+gulaPasir()+butter()+ragi()+susuBubuk()+susuCair() + esBatu() + telur()  );
         return totalBerat;
     }
